pyproject.toml
[project]
name = "debian-cloud-init"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
